+++ next-layout.d.ts
@@ -1,0 +1,13 @@
+export type NextPageWithLayout = NextPage & {
+  getLayout: (page: React.ReactNode) => void
+  allowedRoles?: string[]
+}
+
+export type NextComponentWithLayout = NextComponentType & {
+  getLayout: (page: React.ReactNode) => React.ReactNode
+  allowedRoles?: string[]
+}
+
+export type AppPropsWithLayout = AppProps & {
+  Component: NextComponentWithLayout
+}
