+++ src/pages/event/[pid]/edit.tsx
@@ -1,9 +1,11 @@
 import EventForm from '@/features/event/components/EventForm'
 import { GET_EVENT_DETAILS } from '@/graphql/event'
+import { UserRole } from '@/types/membership'
 import { useQuery } from '@apollo/client'
 import { useRouter } from 'next/router'
+import { ReactNode } from 'react'
 
-const EditEvent = () => {
+const EditEventPage = () => {
   const router = useRouter()
   const eventId = router.query.pid
   const { loading, data } = useQuery(GET_EVENT_DETAILS, {
@@ -19,4 +21,10 @@ const EditEvent = () => {
   return <EventForm event={data.eventDetails} />
 }
 
-export default EditEvent
+EditEventPage.getLayout = function getLayout(page: ReactNode) {
+  return <>{page}</>
+}
+
+EditEventPage.allowedRoles = [UserRole.ADMIN]
+
+export default EditEventPage
