+++ src/pages/event/add.tsx
@@ -1,13 +1,13 @@
 import EventForm from '@/features/event/components/EventForm'
-import { CREATE_EVENT_MUTATION, GET_EVENTS_QUERY } from '@/graphql/event'
-import { useToasts } from '@/hooks/useToasts'
-import { useMutation } from '@apollo/client'
-import Router, { useRouter } from 'next/router'
-import React from 'react'
-import { SubmitHandler, useForm } from 'react-hook-form'
+import { UserRole } from '@/types/membership'
+import React, { ReactNode } from 'react'
 
-const AddEvent = () => {
+const AddEventPage = () => {
   return <EventForm />
 }
 
-export default AddEvent
+AddEventPage.getLayout = function getLayout(page: ReactNode) {
+  return <>{page}</>
+}
+
+export default AddEventPage
